requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
tortoise-orm[aiosqlite]